Angular ngIf Directive and the Elvis Operator
learned that ngIf is a useful tool to hide and/or show elements based on a boolean. syntax also allows for an else statement to load in, for instance, alternate templates if item is not present. Can be used to ensure uniformity of data and protect against malformed or erroneous data.